+++ security/dvs_reactor.go
@@ -21,6 +21,10 @@ import (
 	"github.com/0xPellNetwork/pelldvs/types"
 )
 
+const (
+	responseDigestLenLimit = 32
+)
+
 type DVSReactor struct {
 	config            config.PellConfig
 	ProxyApp          proxy.AppConns
@@ -147,9 +151,13 @@ func (dvs *DVSReactor) OnRequest(request avsitypes.DVSRequest) (*avsitypes.DVSRe
 		Request:  &request,
 		Operator: operators,
 	})
-
 	if err != nil {
 		return nil, err
+	}
+
+	// Check if responseDigest length is equal to 32
+	if len(responseProcessDVSRequest.ResponseDigest) != responseDigestLenLimit {
+		return nil, fmt.Errorf("responseDigest length %d is not equal to %d", responseProcessDVSRequest.ResponseDigest, responseDigestLenLimit)
 	}
 
 	reqResIdx := avsitypes.DVSRequestResult{
@@ -169,12 +177,9 @@ func (dvs *DVSReactor) OnRequest(request avsitypes.DVSRequest) (*avsitypes.DVSRe
 	}
 	dvs.logger.Debug("responseWithSignature", "signature", signature)
 
-	var digestArr [32]byte
-	copy(digestArr[:], responseProcessDVSRequest.ResponseDigest)
 	g1p := bls.G1Point{
 		G1Affine: signature.G1Affine,
 	}
-
 	sig := bls.Signature{G1Point: &g1p}
 
 	responseWithSingature := aggtypes.ResponseWithSignature{
@@ -182,7 +187,7 @@ func (dvs *DVSReactor) OnRequest(request avsitypes.DVSRequest) (*avsitypes.DVSRe
 		Signature:   &sig,
 		OperatorID:  dvs.dvsState.operatorID,
 		RequestData: request,
-		Digest:      digestArr,
+		Digest:      [32]byte(responseProcessDVSRequest.ResponseDigest),
 	}
 
 	// Create a channel to receive validated response