+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 # CHANGELOG
 
-## v0.2.0
+## [Unreleased v0.3.0]
 
-*March 1, 2025*
+### Features
+
+### Improvements
+
+(CI) [#7](https://github.com/0xPellNetwork/pelldvs/pull/7) ci: add changelog check in CI  
+
+(CI) [#11](https://github.com/0xPellNetwork/pelldvs/pull/11) ci: retrieve E2E port config from GitHub variables  
+
+(refactor) [#9](https://github.com/0xPellNetwork/pelldvs/pull/9) refactor: rename abci to avsi  
+
+### Bug Fixes
+
+(dvs) [#6](https://github.com/0xPellNetwork/pelldvs/pull/6) fix: verify DVS request response digest length is 32 
+
+## v0.2.2
+
+v0.2.2 includes a version release CI script fix.  
+
+## v0.2.0
 
 This release marks the transition from a previously private repository to the public release. After a period of internal development, the repository is now archived and versioned as v0.2.0.